pyproject.toml
[project]
name = "demo-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.16"
dependencies = [
    "fastapi[all]>=0.115.13",
    "pandas>=2.2.3",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-cov>=4.1.0",
    "uvicorn>=0.35.0",
    "xgboost>=2.0.3",
    "scikit-learn>=1.4.2",
    "seaborn>=0.13.2",
    "httpx>=0.27.0",
    "python-dotenv>=1.1.1",
    "dotenv>=0.9.9",
    "google-cloud-storage>=3.2.0",
    "locust>=2.38.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
